Implement batching for broadcast challenge
This significantly reduces the number of messages per operation, but
latency remains high. I think our linear topology is the issue because
adding nodes increases the latency. Logically this makes sense if you
consider the number of hops it takes for a message from the first node
to reach the last node.