Working implementation for challenge 4
Since the key-value we're relying on is only sequentially consistent, we
need to perform a noop compare and swap (CAS) that leaves the value the
same, but updates the ID. By doing this, we ensure the write persists
and we aren't reading stale data.

This is a working implementation, but I would like to cleanup the code
to be more generic and ergonomic.